Add files via upload
These are the main functions of the Generative Single Shot Tracking. They include SAMPLE_IMAGE, PREDICT_BB, BoW_Update, and GO. Sampling the image produces the image region that will be used to predict the BB size and location. Predicting BB uses a 2D gaussian fir and Kalman Filter to predict and output the estimated BB size and location. The BoW is then updated with the predicted BB region. The genetic optimization uses an iterative approach to tune the hyper-parameters.